requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
langchain==0.1.4
langchain-google-genai==0.0.6
langchain-community==0.0.16
chromadb==0.4.22
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0
streamlit==1.28.1
plotly==5.18.0
pandas==2.1.4
requests